Remove isolated voxels at the edge of the output segmentation (#2437)
* sct_deepseg_sc: cosmetic

* deepseg_sc.core: Resampling Image instead of file

To minimize i/o. Also renamed some variables for clarification (unfinished)

* deepseg_sc/core: Renamed variabled according to convention

* deepseg_sc/postprocessing: New file dealing with postproc code

* deepseg_sc/postprocessing: Moved postproc code here

* deepseg_sc/core: Moved postproc code to postproc

* deepseg_sc.core: Updated input variable

* deepseg_sc.postprocessing: Implemented check for edge outliers

* deepseg_sc.postprocessing: Fixed logger format

* deep_sc/postprocessing: add post_processing_slice_wise

* deep_sc/postprocessing: logger.warning instead of logger.info

* deep_sc/core: remove i/o of im_labels when viewer

* deep_sc/core: resample im_viewer without i/o

* deep_lesion/core: resample im_viewer without i/o

* unit_testing/test_deepseg_sc.py: correct number of output for find_centerline

* deepseg_lesion: define im_viewer
