+++ spinalcordtoolbox/deepseg_sc/core.py
@@ -6,12 +6,13 @@ import os, sys, logging
 
 import numpy as np
 from scipy.ndimage.measurements import center_of_mass, label
-from scipy.ndimage.morphology import binary_fill_holes
 from skimage.exposure import rescale_intensity
 from scipy.ndimage import distance_transform_edt
+import nibabel as nib
 
 from spinalcordtoolbox import resampling
-from spinalcordtoolbox.deepseg_sc.cnn_models import nn_architecture_seg, nn_architecture_ctr
+from .cnn_models import nn_architecture_seg, nn_architecture_ctr
+from .postprocessing import post_processing_volume_wise, post_processing_slice_wise
 from spinalcordtoolbox.image import Image, empty_like, change_type, zeros_like
 from spinalcordtoolbox.centerline.core import ParamCenterline, get_centerline, _call_viewer_centerline
 
@@ -113,9 +114,7 @@ def find_centerline(algo, image_fname, contrast_type, brain_bool, folder_output,
         im_labels = _call_viewer_centerline(Image(image_fname))
         im_centerline, arr_centerline, _, _ = get_centerline(im_labels, param=ParamCenterline())
         centerline_filename = sct.add_suffix(image_fname, "_ctr")
-        labels_filename = sct.add_suffix(image_fname, "_labels-centerline")
         im_centerline.save(centerline_filename)
-        im_labels.save(labels_filename)
 
     elif algo == 'file':
         centerline_filename = sct.add_suffix(image_fname, "_ctr")
@@ -140,7 +139,10 @@ def find_centerline(algo, image_fname, contrast_type, brain_bool, folder_output,
         im_split_lst = split_data(Image(centerline_filename), dim=2)
         im_split_lst[0].save(centerline_filename)
 
-    return fname_res, centerline_filename
+    if algo != 'viewer':
+        im_labels = None
+
+    return fname_res, centerline_filename, im_labels
 
 
 def scale_intensity(data, out_min=0, out_max=255):
@@ -199,104 +201,6 @@ def crop_image_around_centerline(im_in, ctr_in, crop_size):
 
     im_new.data = data_im_new
     return x_lst, y_lst, z_lst, im_new
-
-
-def _remove_extrem_holes(z_lst, end_z, start_z=0):
-    """Remove extrem holes from the list of holes so that we will not interpolate on the extrem slices."""
-    if start_z in z_lst:
-        while start_z in z_lst:
-            z_lst = z_lst[1:]
-            start_z += 1
-        if len(z_lst):
-            z_lst.pop(0)
-
-    if end_z in z_lst:
-        while end_z in z_lst:
-            z_lst = z_lst[:-1]
-            end_z -= 1
-
-    return z_lst
-
-
-def _list2range(lst):
-    tmplst = lst[:]
-    tmplst.sort()
-    start = tmplst[0]
-
-    currentrange = [start, start + 1]
-
-    for item in tmplst[1:]:
-        if currentrange[1] == item:  # contiguous
-            currentrange[1] += 1
-        else:  # new range start
-            yield list(currentrange)
-            currentrange = [item, item + 1]
-
-    yield list(currentrange)  # last range
-
-
-def _fill_z_holes(zz_lst, data, z_spaccing):
-    data_interpol = np.copy(data)
-
-    for z_hole_start, z_hole_end in list(_list2range(zz_lst)):
-        z_ref_start, z_ref_end = z_hole_start - 1, z_hole_end
-        slice_ref_start, slice_ref_end = data[:, :, z_ref_start], data[:, :, z_ref_end]
-
-        hole_cur_lst = list(range(z_hole_start, z_hole_end))
-        lenght_hole = len(hole_cur_lst) + 1
-        phys_lenght_hole = lenght_hole * z_spaccing
-
-        denom_interpolation = (lenght_hole + 1)
-
-        if phys_lenght_hole < 10:
-            logger.warning('Filling a hole in the segmentation around z_slice #:' + str(z_ref_start))
-
-            for idx_z, z_hole_cur in enumerate(hole_cur_lst):
-                num_interpolation = (lenght_hole - idx_z - 1) * slice_ref_start  # Contribution of the bottom ref slice
-                num_interpolation += (idx_z + 1) * slice_ref_end  # Contribution of the top ref slice
-
-                slice_interpolation = num_interpolation * 1. / denom_interpolation
-                slice_interpolation = (slice_interpolation > 0).astype(np.int)
-
-                data_interpol[:, :, z_hole_cur] = slice_interpolation
-
-    return data_interpol
-
-
-def _remove_blobs(data):
-    """Remove false positive blobs, likely occuring in brain sections."""
-    labeled_obj, num_obj = label(data)
-    if num_obj > 1:  # If there is more than one connected object
-        bigger_obj = (labeled_obj == (np.bincount(labeled_obj.flat)[1:].argmax() + 1))
-
-        data2clean = np.copy(data)
-
-        # remove blobs only above the bigger connected object
-        z_max = np.max(np.where(bigger_obj)[2])
-        data2clean[:, :, :z_max + 1] = 0
-
-        labeled_obj2clean, num_obj2clean = label(data2clean)
-        if num_obj2clean:  # If there is connected object above the biffer connected one
-            for obj_id in range(1, num_obj2clean + 1):
-                # if the blob has a volume < 10% of the bigger connected object, then remove it
-                if np.sum(labeled_obj2clean == obj_id) < 0.1 * np.sum(bigger_obj):
-                    sct.printv('Removing small objects above slice#' + str(z_max))
-                    data[np.where(labeled_obj2clean == obj_id)] = 0
-
-    return data
-
-
-def post_processing_volume_wise(im_in):
-    """Post processing function."""
-    data_in = im_in.data.astype(np.int)
-
-    data_in = _remove_blobs(data_in)
-
-    zz_zeros = [zz for zz in range(im_in.dim[2]) if 1 not in list(np.unique(data_in[:, :, zz]))]
-    zz_holes = _remove_extrem_holes(zz_zeros, im_in.dim[2] - 1, 0)
-    # filling z_holes, i.e. interpolate for z_slice not segmented
-    im_in.data = _fill_z_holes(zz_holes, data_in, im_in.dim[6]) if len(zz_holes) else data_in
-    return im_in
 
 
 def scan_slice(z_slice, model, mean_train, std_train, coord_lst, patch_shape, z_out_dim):
@@ -515,23 +419,6 @@ def uncrop_image(ref_in, data_crop, x_crop_lst, y_crop_lst, z_crop_lst):
     return seg_unCrop
 
 
-def post_processing_slice_wise(z_slice, x_cOm, y_cOm):
-    """Keep the largest connected obejct per z_slice and fill little holes."""
-    labeled_obj, num_obj = label(z_slice)
-    if num_obj > 1:
-        if x_cOm is None or np.isnan(x_cOm):  # slice 0 or empty slice
-            z_slice = (labeled_obj == (np.bincount(labeled_obj.flat)[1:].argmax() + 1))
-        else:
-            idx_z_minus_1 = np.bincount(labeled_obj.flat)[1:].argmax() + 1
-            for idx in range(1, num_obj + 1):
-                z_idx = labeled_obj == idx
-                if z_idx[int(x_cOm), int(y_cOm)]:
-                    idx_z_minus_1 = idx
-            z_slice = (labeled_obj == idx_z_minus_1)
-
-    return binary_fill_holes(z_slice, structure=np.ones((3, 3))).astype(np.int)
-
-
 def segment_3d(model_fname, contrast_type, im_in):
     """Perform segmentation with 3D convolutions."""
     from spinalcordtoolbox.deepseg_sc.cnn_models_3d import load_trained_model
@@ -600,13 +487,13 @@ def deep_segmentation_spinalcord(im_image, contrast_type, ctr_algo='cnn', ctr_fi
 
     # find the spinal cord centerline - execute OptiC binary
     logger.info("Finding the spinal cord centerline...")
-    fname_res, centerline_filename = find_centerline(algo=ctr_algo,
-                                                     image_fname=fname_orient,
-                                                     contrast_type=contrast_type,
-                                                     brain_bool=brain_bool,
-                                                     folder_output=tmp_folder_path,
-                                                     remove_temp_files=remove_temp_files,
-                                                     centerline_fname=file_ctr)
+    fname_res, centerline_filename, im_labels_viewer = find_centerline(algo=ctr_algo,
+                                                                        image_fname=fname_orient,
+                                                                        contrast_type=contrast_type,
+                                                                        brain_bool=brain_bool,
+                                                                        folder_output=tmp_folder_path,
+                                                                        remove_temp_files=remove_temp_files,
+                                                                        centerline_fname=file_ctr)
 
     im_nii, ctr_nii = Image(fname_res), Image(centerline_filename)
 
@@ -628,66 +515,70 @@ def deep_segmentation_spinalcord(im_image, contrast_type, ctr_algo='cnn', ctr_fi
         logger.info("Segmenting the spinal cord using deep learning on 2D patches...")
         segmentation_model_fname = \
             os.path.join(sct.__sct_dir__, 'data', 'deepseg_sc_models', '{}_sc.h5'.format(contrast_type))
-        seg_crop_data = segment_2d(model_fname=segmentation_model_fname,
-                                   contrast_type=contrast_type,
-                                   input_size=(crop_size, crop_size),
-                                   im_in=im_norm_in)
+        seg_crop = segment_2d(model_fname=segmentation_model_fname,
+                                 contrast_type=contrast_type,
+                                 input_size=(crop_size, crop_size),
+                                 im_in=im_norm_in)
     elif kernel_size == '3d':
         # segment data using 3D convolutions
         logger.info("Segmenting the spinal cord using deep learning on 3D patches...")
         segmentation_model_fname = \
             os.path.join(sct.__sct_dir__, 'data', 'deepseg_sc_models', '{}_sc_3D.h5'.format(contrast_type))
-        seg_crop_data = segment_3d(model_fname=segmentation_model_fname,
-                                   contrast_type=contrast_type,
-                                   im_in=im_norm_in)
+        seg_crop = segment_3d(model_fname=segmentation_model_fname,
+                                 contrast_type=contrast_type,
+                                 im_in=im_norm_in)
     del im_norm_in
 
     # reconstruct the segmentation from the crop data
     logger.info("Reassembling the image...")
-    seg_uncrop_nii = uncrop_image(ref_in=im_nii,
-                                  data_crop=seg_crop_data,
-                                  x_crop_lst=X_CROP_LST,
-                                  y_crop_lst=Y_CROP_LST,
-                                  z_crop_lst=Z_CROP_LST)
-    fname_res_seg = sct.add_suffix(fname_res, '_seg')
-    seg_uncrop_nii.save(fname_res_seg)
-    del seg_crop_data
+    im_seg = uncrop_image(ref_in=im_nii,
+                          data_crop=seg_crop,
+                          x_crop_lst=X_CROP_LST,
+                          y_crop_lst=Y_CROP_LST,
+                          z_crop_lst=Z_CROP_LST)
+    # fname_res_seg = sct.add_suffix(fname_res, '_seg')
+    # seg_uncrop_nii.save(fname_res_seg)  # for debugging
+    del seg_crop
 
     # resample to initial resolution
-    logger.info("Resampling the segmentation to the original image resolution...")
-    initial_resolution = 'x'.join([str(input_resolution[0]), str(input_resolution[1]), str(input_resolution[2])])
-    fname_res_seg_downsamp = sct.add_suffix(fname_res_seg, '_downsamp')
-
-    resampling.resample_file(fname_res_seg, fname_res_seg_downsamp, initial_resolution,
-                                                           'mm', 'linear', verbose=0)
-    im_image_res_seg_downsamp = Image(fname_res_seg_downsamp)
+    logger.info("Resampling the segmentation to the native image resolution using linear interpolation...")
+    # create nibabel object
+    seg_uncrop_nii_nib = nib.nifti1.Nifti1Image(im_seg.data, im_seg.hdr.get_best_affine())
+    seg_uncrop_nii_nibr = resampling.resample_nib(seg_uncrop_nii_nib, new_size=input_resolution, new_size_type='mm',
+                                                  interpolation='linear')
+    # Convert back to Image type
+    im_seg_r = Image(seg_uncrop_nii_nibr.get_data(), hdr=seg_uncrop_nii_nibr.header, orientation='RPI',
+                     dim=seg_uncrop_nii_nibr.header.get_data_shape())
 
     if ctr_algo == 'viewer':  # resample and reorient the viewer labels
-        fname_res_labels = sct.add_suffix(fname_orient, '_labels-centerline')
-        resampling.resample_file(fname_res_labels, fname_res_labels, initial_resolution, 'mm', 'linear', verbose=0)
-        im_image_res_labels_downsamp = Image(fname_res_labels).change_orientation(original_orientation)
+        im_labels_viewer_nib = nib.nifti1.Nifti1Image(im_labels_viewer.data, im_labels_viewer.hdr.get_best_affine())
+        im_viewer_r_nib = resampling.resample_nib(im_labels_viewer_nib, new_size=input_resolution, new_size_type='mm',
+                                                    interpolation='linear')
+        im_viewer = Image(im_viewer_r_nib.get_data(), hdr=im_viewer_r_nib.header, orientation='RPI',
+                            dim=im_viewer_r_nib.header.get_data_shape()).change_orientation(original_orientation)
     else:
-        im_image_res_labels_downsamp = None
+        im_viewer = None
 
-    if verbose == 2:
-        fname_res_ctr = sct.add_suffix(fname_orient, '_ctr')
-        resampling.resample_file(fname_res_ctr, fname_res_ctr, initial_resolution, 'mm', 'linear', verbose=0)
-        im_image_res_ctr_downsamp = Image(fname_res_ctr).change_orientation(original_orientation)
-    else:
-        im_image_res_ctr_downsamp = None
+    # TODO: Deal with that later-- ideally this file should be written when debugging, not with verbose=2
+    # if verbose == 2:
+    #     fname_res_ctr = sct.add_suffix(fname_orient, '_ctr')
+    #     resampling.resample_file(fname_res_ctr, fname_res_ctr, initial_resolution, 'mm', 'linear', verbose=0)
+    #     im_image_res_ctr_downsamp = Image(fname_res_ctr).change_orientation(original_orientation)
+    # else:
+    im_image_res_ctr_downsamp = None
 
-    # binarize the resampled image to remove interpolation effects
-    logger.info("Binarizing the segmentation to avoid interpolation effects...")
+    # Binarize the resampled image to remove interpolation effects
+    logger.info("Binarizing the resampled segmentation...")
     thr = 0.0001 if contrast_type in ['t1', 'dwi'] else 0.5
     # TODO: optimize speed --> np.where is slow
-    im_image_res_seg_downsamp.data[np.where(im_image_res_seg_downsamp.data >= thr)] = 1
-    im_image_res_seg_downsamp.data[np.where(im_image_res_seg_downsamp.data < thr)] = 0
+    im_seg_r.data[np.where(im_seg_r.data >= thr)] = 1
+    im_seg_r.data[np.where(im_seg_r.data < thr)] = 0
 
     # post processing step to z_regularized
-    im_image_res_seg_downsamp_postproc = post_processing_volume_wise(im_in=im_image_res_seg_downsamp)
+    im_seg_r_postproc = post_processing_volume_wise(im_seg_r)
 
     # change data type
-    im_image_res_seg_downsamp_postproc.change_type(np.uint8)
+    im_seg_r_postproc.change_type(np.uint8)
 
     tmp_folder.chdir_undo()
 
@@ -697,8 +588,8 @@ def deep_segmentation_spinalcord(im_image, contrast_type, ctr_algo='cnn', ctr_fi
         tmp_folder.cleanup()
 
     # reorient to initial orientation
-    return im_image_res_seg_downsamp_postproc.change_orientation(original_orientation), \
+    return im_seg_r_postproc.change_orientation(original_orientation), \
            im_nii, \
-           seg_uncrop_nii.change_orientation('RPI'), \
-           im_image_res_labels_downsamp, \
+           im_seg.change_orientation('RPI'), \
+           im_viewer, \
            im_image_res_ctr_downsamp
